urxvt/plugin: don't manage plugin with a submodule
Using a submodule for the URxvt's Perl library limits me to a single
repository.
This is a problem since I want to use the plugin url-picker which is not
in the repository muennich/urxvt-perls.

So instead of using a submodule, let's manage these by myself, anyway
they are rarely updated these days.