emacs/package: install origami
In my quest to Vim-like folding (using markers) I settled on
outline+outshine  (for lack of better alternative) but after
playing/hacking a bit with origami I got something much closer to my Vim
workflow.

One of the biggest advantage is that I can use the same syntax for both
Vim and Emacs (wasn't the case when using outline+outshine).

That being said, a pain point of origami though is that to open/close a
fold you have to be exactly on a fold marker (otherwise it doesn't
work). It's pretty annoying.
Note that there is a special function to avoid this issue when toggling:
`origami-forward-toggle-node` (which is like `origami-toggle-node`, but
more convenient).

I applied this principle to the open and close functions, and voilà! It
works like a charm :)

I also wrote a function to toggle recursively, because the existing one
(`origami-forward-recursively-toggle-node`) is more like a `cycle` than
a `toggle`.